cleanup code

+++ src/files.cpp
@@ -14,7 +14,7 @@ std::string getexepath()
 {
     char result[ PATH_MAX ];
     ssize_t count = readlink( "/proc/self/exe", result, PATH_MAX );
-    return std::string( result, (count > 0) ? count : 0 );
+    return std::string(result, static_cast<unsigned long>((count > 0) ? count : 0));
 }
 
 void File::write(std::string input) {
